IMU code implemented with mutex objects and std::thread
