+++ robot/src/rt/test_imu/rt_imu.c
@@ -13,7 +13,7 @@
 #include <math.h>
 #include <pthread.h>
 
-#include <endian.h>
+#include <machine/endian.h>
 #include <stdint.h>
 
 #include "config.h"\
